refactor: split regulations service in many files
to improve readability
