+++ app/helpers/regulations_utils.py
@@ -1,0 +1,9 @@
+from collections import namedtuple
+
+DAY = 86400
+HOUR = 3600
+MINUTE = 60
+
+ComputationResult = namedtuple(
+    "ComputationResult", ["success", "extra"], defaults=(False, None)
+)
